requirements.txt.txt
python-telegram-bot
google-generativeai
rapidfuzz
